Extract new-releases icon from MaterialIcon
